Quick save

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import { Component, Input } from '@angular/core';
+import { Component, ElementRef, Input, OnInit, ViewChild } from '@angular/core';
 import { FormGroup } from '@angular/forms';
 import { BasketService } from 'src/app/basket/basket.service';
 import { Basket } from 'src/app/models/basket';
@@ -6,28 +6,83 @@ import { CheckoutService } from '../checkout.service';
 import { ToastrService } from 'ngx-toastr';
 import { Address } from 'src/app/models/user';
 import { NavigationExtras, Route, Router } from '@angular/router';
+import { Stripe, StripeCardCvcElement, StripeCardExpiryElement, StripeCardNumberElement, loadStripe } from '@stripe/stripe-js';
 
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent {
+export class CheckoutPaymentComponent implements OnInit {
   @Input() checkoutForm?: FormGroup;
+  @ViewChild('cardNumber') cardNumberElement?: ElementRef;
+  @ViewChild('cardExpiry') cardExpiryElement?: ElementRef;
+  @ViewChild('cardCvc') cardCvcElement?: ElementRef;
+  stripe: Stripe | null = null;
+  cardNumber?: StripeCardNumberElement;
+  cardExpiry?: StripeCardExpiryElement;
+  cardCvc?: StripeCardCvcElement;
+  cardErrors: any;
 
-  constructor(private basketService: BasketService, private checkoutService: CheckoutService, private toastr: ToastrService, private router: Router){}
 
-  submitOrder(){
+
+  constructor(private basketService: BasketService, private checkoutService: CheckoutService, private toastr: ToastrService, private router: Router) { }
+  ngOnInit(): void {
+    loadStripe('pk_test_51OCMSTBwFRbTh8QWkQ99jc0NFcENcFW85VVKZer6NOyogNehz7XqS7KilpVSmMSLquP3gxjAJ2PBcmtDGuUDNy9k00crRTdwba').then(stripe => {
+      this.stripe = stripe;
+      const elements = stripe?.elements();
+      if (elements) {
+        this.cardNumber = elements.create('cardNumber');
+        this.cardNumber.mount(this.cardNumberElement?.nativeElement);
+        this.cardNumber.on('change', event => {
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+
+        this.cardExpiry = elements.create('cardExpiry');
+        this.cardExpiry.mount(this.cardExpiryElement?.nativeElement);
+        this.cardExpiry.on('change', event => {
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+
+        this.cardCvc = elements.create('cardCvc');
+        this.cardCvc.mount(this.cardCvcElement?.nativeElement);
+        this.cardCvc.on('change', event => {
+          if (event.error) this.cardErrors = event.error.message;
+          else this.cardErrors = null;
+        })
+      }
+    })
+  }
+
+
+
+  submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
-    if(!basket) return;
+    if (!basket) return;
     const orderToCreate = this.getOrderToCreate(basket)
-    if(!orderToCreate) return;
+    if (!orderToCreate) return;
     this.checkoutService.createOrder(orderToCreate).subscribe({
       next: order => {
         this.toastr.success('Order created successfully');
-        this.basketService.deleteLocalBasket();
-        const navigationExtras: NavigationExtras = {state: order};
-        this.router.navigate(['checkout/success'], navigationExtras)
+        this.stripe?.confirmCardPayment(basket.clientSecret!, {
+          payment_method: {
+            card: this.cardNumber!,
+            billing_details: {
+              name: this.checkoutForm?.get('paymentForm')?.get('nameOnCard')?.value
+            }
+          }
+        }).then(result => {
+          console.log(result);
+          if (result.paymentIntent) {
+            this.basketService.deleteLocalBasket();
+            const navigationExtras: NavigationExtras = { state: order };
+            this.router.navigate(['checkout/success'], navigationExtras)
+          } else{
+            this.toastr.error(result.error.message);
+          }
+        })
       }
     })
   }
@@ -35,11 +90,10 @@ export class CheckoutPaymentComponent {
     const deliveryMethodId = this.checkoutForm?.get('deliveryForm')?.get('deliveryMethod')?.value;
     const shipToAddress = this.checkoutForm?.get('addressForm')?.value as Address;
     if (!deliveryMethodId || !shipToAddress) return;
-    return{
+    return {
       basketId: basket.id,
       deliveryMethodId: deliveryMethodId,
       shipToAddress: shipToAddress
     }
   }
 }
-
